requirements.txt
Flask==3.0.0
flask-cors==4.0.0
google-generativeai==0.8.6
python-dotenv==1.2.2
gunicorn==21.2.0
